requirements.txt
pandas
scipy
streamlit
plotly_express
